pyproject.toml
[project]
name = "option-research"
version = "0.1.0"
description = "Directional score model for choosing call vs put trades"
authors = [{name = "Option Research"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "supabase>=2.0.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "python-dateutil>=2.8.0",
    "finnhub-python>=2.4.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["lib"]
